Get rid of input_mouse_grabbed - move retroarch types to retroarch_types.h
