+++ core.h
@@ -23,129 +23,9 @@
 
 #include <retro_common_api.h>
 
-#include "core_type.h"
-#include "input/input_defines.h"
-#include "disk_control_interface.h"
+#include "retroarch_types.h"
 
 RETRO_BEGIN_DECLS
-
-enum
-{
-   /* Polling is performed before
-    * call to retro_run. */
-   POLL_TYPE_EARLY = 0,
-
-   /* Polling is performed when requested. */
-   POLL_TYPE_NORMAL,
-
-   /* Polling is performed on first call to
-    * retro_input_state per frame. */
-   POLL_TYPE_LATE
-};
-
-typedef struct rarch_memory_descriptor
-{
-   struct retro_memory_descriptor core;        /* uint64_t alignment */
-   size_t disconnect_mask;
-} rarch_memory_descriptor_t;
-
-typedef struct rarch_memory_map
-{
-   rarch_memory_descriptor_t *descriptors;
-   unsigned num_descriptors;
-} rarch_memory_map_t;
-
-typedef struct rarch_system_info
-{
-   struct retro_location_callback location_cb; /* ptr alignment */
-   disk_control_interface_t disk_control;      /* ptr alignment */
-   struct retro_system_info info;              /* ptr alignment */
-   rarch_memory_map_t mmaps;                   /* ptr alignment */
-   const char *input_desc_btn[MAX_USERS][RARCH_FIRST_META_KEY];
-   struct
-   {
-      struct retro_subsystem_info *data;
-      unsigned size;
-   } subsystem;
-   struct
-   {
-      struct retro_controller_info *data;
-      unsigned size;
-   } ports;
-   unsigned rotation;
-   unsigned performance_level;
-   char valid_extensions[255];
-   bool load_no_content;
-   bool supports_vfs;
-} rarch_system_info_t;
-
-typedef struct retro_ctx_input_state_info
-{
-   retro_input_state_t cb;
-} retro_ctx_input_state_info_t;
-
-typedef struct retro_ctx_cheat_info
-{
-   const char *code;
-   unsigned index;
-   bool enabled;
-} retro_ctx_cheat_info_t;
-
-typedef struct retro_ctx_api_info
-{
-   unsigned version;
-} retro_ctx_api_info_t;
-
-typedef struct retro_ctx_region_info
-{
-  unsigned region;
-} retro_ctx_region_info_t;
-
-typedef struct retro_ctx_controller_info
-{
-   unsigned port;
-   unsigned device;
-} retro_ctx_controller_info_t;
-
-typedef struct retro_ctx_memory_info
-{
-   void *data;
-   size_t size;
-   unsigned id;
-} retro_ctx_memory_info_t;
-
-typedef struct retro_ctx_load_content_info
-{
-   struct retro_game_info *info;
-   const struct string_list *content;
-   const struct retro_subsystem_info *special;
-} retro_ctx_load_content_info_t;
-
-typedef struct retro_ctx_serialize_info
-{
-   const void *data_const;
-   void *data;
-   size_t size;
-} retro_ctx_serialize_info_t;
-
-typedef struct retro_ctx_size_info
-{
-   size_t size;
-} retro_ctx_size_info_t;
-
-typedef struct retro_ctx_environ_info
-{
-   retro_environment_t env;
-} retro_ctx_environ_info_t;
-
-typedef struct retro_callbacks
-{
-   retro_video_refresh_t frame_cb;
-   retro_audio_sample_t sample_cb;
-   retro_audio_sample_batch_t sample_batch_cb;
-   retro_input_state_t state_cb;
-   retro_input_poll_t poll_cb;
-} retro_callbacks_t;
 
 #ifdef HAVE_REWIND
 bool core_set_rewind_callbacks(void);
