+++ gfx/common/win32_common.c
@@ -1053,11 +1053,11 @@ static LRESULT CALLBACK wnd_proc_common_internal(HWND hwnd,
          break;
 #ifdef HAVE_CLIP_WINDOW
       case WM_SETFOCUS:
-         if (input_mouse_grabbed())
+         if (input_state_get_ptr()->grab_mouse_state)
             win32_clip_window(true);
          break;
       case WM_KILLFOCUS:
-         if (input_mouse_grabbed())
+         if (input_state_get_ptr()->grab_mouse_state)
             win32_clip_window(false);
          break;
 #endif
@@ -1128,7 +1128,7 @@ static LRESULT CALLBACK wnd_proc_winraw_common_internal(HWND hwnd,
          break;
       case WM_SETFOCUS:
 #ifdef HAVE_CLIP_WINDOW
-         if (input_mouse_grabbed())
+         if (input_state_get_ptr()->grab_mouse_state)
             win32_clip_window(true);
 #endif
 #if !defined(_XBOX)
@@ -1138,7 +1138,7 @@ static LRESULT CALLBACK wnd_proc_winraw_common_internal(HWND hwnd,
          break;
       case WM_KILLFOCUS:
 #ifdef HAVE_CLIP_WINDOW
-         if (input_mouse_grabbed())
+         if (input_state_get_ptr()->grab_mouse_state)
             win32_clip_window(false);
 #endif
 #if !defined(_XBOX)
@@ -1264,11 +1264,11 @@ static LRESULT CALLBACK wnd_proc_common_dinput_internal(HWND hwnd,
          break;
 #ifdef HAVE_CLIP_WINDOW
       case WM_SETFOCUS:
-         if (input_mouse_grabbed())
+         if (input_state_get_ptr()->grab_mouse_state)
             win32_clip_window(true);
          break;
       case WM_KILLFOCUS:
-         if (input_mouse_grabbed())
+         if (input_state_get_ptr()->grab_mouse_state)
             win32_clip_window(false);
          break;
 #endif
