+++ input/input_driver.h
@@ -45,6 +45,10 @@
 #include "../configuration.h"
 #include "../performance_counters.h"
 
+#ifdef HAVE_COMMAND
+#include "../command.h"
+#endif
+
 RETRO_BEGIN_DECLS
 
 /**
@@ -294,19 +298,37 @@ struct rarch_joypad_driver
 
 typedef struct
 {
+   /**
+    * Array of timers, one for each entry in enum input_combo_type.
+    */
+   rarch_timer_t combo_timers[INPUT_COMBO_LAST];
+
    /* pointers */
    input_driver_t                *current_driver;
    void                          *current_data;
    const input_device_driver_t   *primary_joypad;        /* ptr alignment */
    const input_device_driver_t   *secondary_joypad;      /* ptr alignment */
+#ifdef HAVE_COMMAND
+   command_t *command[MAX_CMD_DRIVERS];
+#endif
+#ifdef HAVE_NETWORKGAMEPAD
+   input_remote_t *remote;
+#endif
+
+   turbo_buttons_t turbo_btns; /* int32_t alignment */
+
+   input_mapper_t mapper;          /* uint32_t alignment */
 
    /* primitives */
    bool        nonblocking_flag;
+   bool keyboard_linefeed_enable;
 
-   /**
-    * Array of timers, one for each entry in enum input_combo_type.
-    */
-   rarch_timer_t combo_timers[INPUT_COMBO_LAST];
+   bool block_hotkey;
+   bool block_libretro_input;
+   bool grab_mouse_state;
+   bool analog_requested[MAX_USERS];
+   bool keyboard_mapping_blocked;
+
 } input_driver_state_t;
 
 
@@ -801,8 +823,6 @@ float input_get_sensor_state(unsigned port, unsigned id);
 
 bool input_set_sensor_state(unsigned port,
       enum retro_sensor_action action, unsigned rate);
-
-bool input_mouse_grabbed(void);
 
 void *input_driver_init_wrap(input_driver_t *input, const char *name);
