+++ input/input_remapping.h
@@ -26,18 +26,6 @@
 #include "input_defines.h"
 #include "input_types.h"
 
-typedef struct input_mapper
-{
-   /* Left X, Left Y, Right X, Right Y */
-   int16_t analog_value[MAX_USERS][8];
-   /* The whole keyboard state */
-   uint32_t keys[RETROK_LAST / 32 + 1];
-   /* RetroPad button state of remapped keyboard keys */
-   unsigned key_button[RETROK_LAST];
-   /* This is a bitmask of (1 << key_bind_id). */
-   input_bits_t buttons[MAX_USERS];
-} input_mapper_t;
-
 RETRO_BEGIN_DECLS
 
 /**
